moved selection, pagination, busy from BTableLite to BTable
